services/api/pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rrga-api"
version = "0.1.0"
description = "Regulatory Register Graph API backend"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16.5",
  "fastapi>=0.135.3",
  "httpx>=0.28.1",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "psycopg[binary]>=3.2.12",
  "pydantic-settings>=2.11.0",
  "python-dateutil>=2.9.0.post0",
  "rapidfuzz>=3.14.1",
  "redis>=6.4.0",
  "sqlalchemy>=2.0.44",
  "tenacity>=9.1.2",
  "typer>=0.20.0",
  "uvicorn[standard]>=0.38.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.14.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.setuptools.packages.find]
where = ["."]
